Simulation.fromFile resolves model path
Resolve first arugment of `Simulation.fromFile` to full path. Note that
without the fully resolved path, the onus is on the C++ core to resolve
the path fully and break it into filepath and filename to be used by
library functions such as
https://github.com/ssandrews/Smoldyn/blob/8d3659afb3a0e103c3f40845e11cf734bd176b66/source/Smoldyn/libsmoldyn.cpp#L297:23

This can also be done at C++ level using std::filesystem library from
C++17 standard.