Fixes to failing test.
setOutputFile sets the name of the output file. There is currently no way
to set the full path of output file. The outfile is always generated in the
current working directory. Hence, it is neccessary to run example file
in the its own directory to make sure that file can be read later by passing
the name of the file.

Modified example/CMakeFile.txt for this effect.